vCPU: Don't clobber vAC for SYS calls
Introduction of sysFn, but with limited success as switching
between calls still clobbers vAC (and costs more space and
time now for the additional 'STW sysFn'). The other alternative
is using inline arguments but that becomes horrible to interpret.
So lets see how far we get with this arrangement: Mandelbrot
has become a bit faster, and all other apps seem unharmed.