Use PROGMEM to store program code and pad pages with 0x00 in Loader
Dynamic RAM on the Arduino Uno is limited to 2k, severely limiting the
size of programs that can be sent with the Loader (Snake, for example, is
only just within this size limitation). Use program memory to store
programs, allowing programs to be around 32k.

In addition, loading larger programs (such as Snake) is very unreliable
unless you pad out each page with zeroes, so modify the loader program to
do this.