Ver 0.4.3
- Added a Tetris sample, this is not finished, (still working on animations
and sound), but is completely playable. This is a huge sample that stores
tetrominoes and multiple subroutines within the unused parts of video memory.
- Added strings to DB command, any combination of valid numerics plus strings is
possible.
- Strings are expected to start with a length byte, (256 bytes maximum), if you
expect compatibility with the helper macros and the debugger gprintf command.
- Refactored the tokeniser, it's now smarter, stronger and has been rebuilt...
- Fixed a bug with multiple macros on one file.
- Added the gprintf command to the assembler and debugger; this is an extrememly
powerful logging command that is invivisible to the assembler, but registers
the current vCPU address with a set of variables that may be logged in real
time to stderr. e.g. gprintf("%d %d %s", *scoreScratch, *scoreDelta, *score_string)
will log 3 variables in real time whenever the vCPU hits the vPC address
associated with the gprintf. The asteriks signifies variable indirection and the
syntax closely models that of the standard printf.