Ver 0.5.3
- Added a vCPu usage display, this counts the maximum vCPU instruction slots
available in a frame. There is also a second counter that counts instruction
slots in a frame excluding an exclusion zone marked by a CpuUsageAddressA
and CpuUsageAddressB. These are usable within vCPU ASM code and editable
through the emulators user interface. Typically they are used tro mark a region
of dead code, (e.g. VBlank wait loop).
- Added VideoB, VideoC and VideoBC scanline modes, these are selectable in the
emulator by default using F3. The emulator patches in the relevant code into
the ROMS in real time, these are not faked in anyway, if the same code was
patched into a real Gigatron's ROM, then it would look and behave exactly the
same way as the emulator.
- Refactored some graphics and timing code.
- Updated README.md's, added a README.md in the gcl directory.