BabelFish: Also accept bare CR (\r, 0x0D, ^M) as newline for commands
Using picoterm on OSX sends just '\r' (CR, ^M, 0x0D) when hitting the
enter/return key. (Unless picoterm's output mapping changes it with
"--omap crlf". But that is annoying.)

This isn't picoterm's fault but how the Terminal window behaves in
"raw" mode. Same behavior on Linux. Just let BabelFish accept these.

If both CR and LF are sent, the second will be an empty line for
doCommand() and nothing is done with that.