BabelFish: Put games and demo applications under Ctrl-Fxx keys
Plus an overall cleanup of the configuaration sections,
using more of the Arduino pin mapping functions where possible.
The pin mapping is now shown to the serial console as well
(hope I didn't mess up for the different platforms).

:BabelFish platform=ArduinoUno
:Pins:
: Gigatron data=13 latch=12 pulse=11
: Keyboard clock=3 data=4
: Controller data=-1
:PROGMEM slots:
: 1) Tiny BASIC
: 2) WozMon
: 3) Terminal
: 4) Blinky
: 5) Bricks game [xbx]
: 6) Lines demo [at67]
: 7) Game of Life demo [at67]
: 8) Starfield demo [at67]
: 9) Tetris game [at67]
: 10) (Empty)
: 11) (Empty)
: 12) (Empty)
:Echo off
:Type 'H' for help
:Gigatron OK
Cmd?

The function key number (or slot number) is now also
an optional scenario to the 'P' command:

: P[<n>]   Transfer object file from PROGMEM slot <n> [1..12]

For example:
  Cmd? R
  Cmd? L
  Cmd? P9
to start tetris

The bare function keys (typed without Ctrl) will be sent to the
Gigatron as non-ASCI codes 0xC1..0xCC (F1..F12). There is nothing
on the Gigatron that uses these yet, but that might come.
[These keys needed codes because the C=16 and C=64 keyboards
also can send function key strokes.]