PWM: Go back to 1 bit per frame
Nitpicking LCD monitor doesn't appreciate too much variation.
Frame to frame deltas of +/-2 scanlines are still ok. Deltas
of 3 or more will blank the screen during transmission.

The maximum SAVE time for a small BASIC program into the ATtiny85
EEPROM will will therefore become 511*8 / 60 = 68 seconds, or just
below 1 minute if we decide to go to 7-bits transmissions (but
lets not go there.)

We should consider a 4-bit-dont-be-kind-to-VGA burst mode.