BM9.gtb: Adjusted for Gigatron TinyBASIC v2
TinyBASIC doesn't have floating point, so any comparison is a bit
off.  This version attempts to preserve the spirit of the original
as good as possible.

Note that lines 130 and 140 effectively call the same slow 16-bit
integer division routine twice with the same arguments. The checks
and control flow of the program is preserved.