Add some initial test bits
Including a single test that runs gtemu.c and checks its execution trace
and output. For now, tests must be compiled and run manually. The test
source can be compiled two ways: once for the target machine (with
-DTEST) and once for the host (without -DTEST). The test is then run by
invoking the host program.

These steps will be enshrined in a Makefile in the near future.