Docs/Smallest.gcl: Tiny busy loop for failed romType checks
The smallest vCPU application
-----------------------------

1. Gives identifiable visual feedback:
  - Draws two colored lines (one vertical, one horizontal)
  - Continuously blinks a pixel

2. Is interactive:
  - Shifts the screen when a button is pressed, restores when released
  - Changes line color when this button is an arrow button

3. Behaves consistent and doesn't crash:
  - Is compatible with all ROM versions, past, present and future
  - Doesn't rely on undocumented features
  - Doesn't accidently clobber system variables (*)

4. Is as small as possible
  - Occupies 6 bytes of RAM (3 vCPU instructions)
  - Loads at an non-intrusive memory location

Application
-----------

When a program uses features that are only available from a certain
ROM version, it should check the romType variable before using these.
At least it shouldn't crash, and preferably it gives the user some
feedback, no matter how simple. See the discussion on GT1 file
compatibility in Docs/GT1-files.txt, which suggests blinking a pixel.

This tiny loop can be used for this signaling function. The program
enters it when it encounters an insufficient romType value.