GCL-language.txt: elaborate vCPU instructions
'DD' or 'D' are intended to represent a single byte. 'LL HH' is
intended to represent a 16-bit word in little-endian order. '[X]'
is intended to represent a RAM location as in the native instruction
set. 'vAC=D' is intended to mean that the (16-bits) vAC gets the
8-bit (unsigned) value D, and therefore clearing its high 8 bits
in the process.