Video: add control variable to black out the top of screen
With this we can crop the top of the screen. This can be useful to
free up screen memory for other purposes, but it also speeds up the
system somewhat.

videoStart_DEVROM == 0x01f9 == 505 (decimal)

Valid values are ONLY the even numbers from 0 to 238! Odd numbers
and larger numbers crash the Gigatron or destroy the display.

For a demonstration, see: https://youtu.be/oQWedkM6zE4