BabelFish: shorter CtrlAltDel delay when held down for long time (#47)
Small rewriting of the `hold' logic for the Ctrl-Alt-Del combination.
Now a 2500 ms long [Start] signal will be still generated, and longer
only when hold down longer. This avoids the accidental triggering of
a second reset and, with that, the Easter Egg.

Now people holding down the key combination all the way until the
first reset (and only then release it), won't get surprised by a
second reset 2 seconds later.

Tested OK for ATtiny85.