BabelFish: Loosen criteria for detection of valid /vSync signal
The detectGigatron() function samples for 85 ms, which corresponds
to just over 5 frames. There will be approximately 2656 scanlines
sampled, but this number isn't exact and depends on the granularity
of the AVR's millis() function.

On a bad day, we can sample the vertical pulse one extra time,
registering it 6 times instead of the expected 5. This gives an
overshoot of 6/5 = 120%. With the fail criterium exactly at 120%,
it will often fail.

Therefore relax the criterium to 1.25 instead: The vertical
tolerance becomes 95% to 125%.