BabelFish: Full BASIC program transfer to PC/laptop
This change allows sending the complete BASIC program to your
computer, no matter how large it is.

When saving a program from TinyBASIC or MSBASIC, it gets transmitted
to BabelFish as ASCII text lines. Now it depends on BabelFish's
serial session state what it will do with it:

When BabelFish is in idle mode (indicated by the "Cmd?" prompt),
the program is still saved into the microcontroller's EEPROM area,
exactly as before. If the program doesn't fit, the Gigatron receives
a Ctrl-C as before. The only change is that in this mode there's
no forwarding to PC/laptop any more. You need to be in Terminal
mode for that now.

In Terminal mode (enter with the 'T' command), now the complete
program is echoed to the host. Previously when typine SAVE in this
mode typing SAVE, all was silently ignored...) Note that the program
still isn't saved to the EEPROM in this mode. But you can capture
it from your terminal window and put it in a file on your computer.
Even if it wouldn't have fit in the EEPROM.