Port several files to Python3
This includes all of the files required to build the dev. ROM, but not
any of the files needed for the older roms.

Support for Python 2.7 is retained. Running dev.py in Python3 results
in the same ROM image as when running in Python 2, which is itself
unchanged from before these changes.

The changes are mostly print and division related, although there are
some bytes/string changes too.