asm.py: make real listing and reduce need for C('comment') #127
The dissassembly file can show comments when they are injected with
the C() function. But this is a bit ugly: it is harder to read in
the .py source than regular comments, while regular annotations
don't show up in the .asm file. This leads to two views on the code.
That isn't helpful.

A much better solution is to produce an old-fashioned listing of
the source code, along with the assembly result on the left. This
is possible with the Python inspect module!

Change: Replace .asm output files with .lst files.

For dev.rom, we now have the Python source lines listed in dev.lst.
Almost all C(xxx) calls could be eliminated between enableListing()
and disableListing().

The justification for the explicit enable/disable (instead of "always
on") is that introspection is rather slow. >90% of the ROM file is
storage area generated from the same macro. With that, looking up
the line number for each word in the storage area, and then deciding
that it was already listed, is very slow..

(Note: This can be remedied with more refactoring that eliminates
the _linenos[] list. Maybe later...)

The transition is quite a bit of work, so for the older ROM files
we don't change the output (for now).

Snippet from dev.lst:
-----------------------------------------------------------------------------------
              0b2d fc32  bra  .sysSb#38   3062  bra('.sysSb#38')                #36
              0b2e de00  st   [y,x++]     3063  st([Y,Xpp])                     #37
                                          3064  label('.sysSb#35')
.sysSb#35:    0b2f 0200  nop              3065  wait(38-35)                     #35
              0b30 0200  nop
              0b31 0200  nop
                                          3066  label('.sysSb#38')
.sysSb#38:    0b32 0124  ld   [$24]       3067  ld([sysArgs+0])                 #38
              0b33 2001  anda $01         3068  anda(1)                         #39
-----------------------------------------------------------------------------------