lcc: Proper argument range for SYS (28...284, not 0..255)
__syscall() was taking the raw operand value for SYS as argument.
That is pretty much useless, because it leads source code that is
difficult to audit, and to bugs that are hard to find with testing
without suitably sensitive VGA monitor (most emulators are not good
in this regard either).

The propoer way is to let the compiler/assembler take the even
number of cycles (ticks), and convert it to excess ticks that vCPU
works with.

The GCL compiler and gt1dump already do the conversion. Now LCC
joins the club.