Apple-1,MSBASIC: Preserve top pixel line
In these programs the video indirection table is setup for a 7-line
text display, with several shared empty pixel lines in between.
However SYS_VDrawBits_134 is still partially overwriting the next
text row because it always plots 8 pixels, not 7. This has the
effect that (usually) the screen's effective top pixel row gets
cleared, and the top text row deprived of its upper pixels.

Workaround this by peeking at the 8th pixel line before replotting
these. This must be done both in Newline and in DrawChar. Although
the code is somewhat clumsy, aesthetically this is better.