Apple-1: Add MASTERMIND
WRITTEN BY STEVE WOZNIAK
PUBLISHED IN DOCTOR DOBBS JOURNAL SEPT 1976

"""MASTERMIND is a number guessing game. In this version the computer
generates a 5-digit, random number where the digits are limited to
0 through 7. The user attemtps to guess the number in the fewest
possible tries. After each guess the computer informs the user as
to how many digits were correct (contained in the random number)
and in the correct position by printing `+' signs, and how many
digits are correct and in the wrong position with `-' signs. The
user is not informed as to exactly which digit positions the `+'
and the `-' signs correspond. A skilled player can usuualle win in
6 or 7 tries."""

https://www.applefritter.com/files/Apple1WozDrDobbsMastermind.pdf