Ver 0.9.3B, (Beta)
Added the 'STRCMP" command to the BASIC compiler for the comparing of 2 strings.
Refactored static initialisation code so that functions that don't make sense during static initialisation commands, (DIM and DEF), will produce warnings.
Fixed functions that were supposed to be working with static initialisation, but weren't, (i.e LEN, ASC, RND, etc).
Upgraded the FONTS sample in the graphics directory to be a little more dynamic.
Added more test code to the test directory.