Loader: Fix channelMask bug introduced in ROM v5a #88
4-channel mode wasn't active after loading.
The root causeh is the rewrite of 35ef40a55e6407f58811f923e5600680139bebc2
(Loader: Improve channelMask handling #88)

Although the address logic for detecting channel overwrites is
correct, we wrongly apply it to the last fragment that BabelFish
sends. Only in this last fragment N=0 and the sent address is the
execution address (typically $200) instead of a fragment address.
From this frame Loader incorrectly deducted that address $2ff would
be overwritten (completely ignoring the N=0). This wrongly triggers
the mechanism that clears the channelMask bits.

The solution is to clear these bits only after the N>0 branch is made.