Fix a few bugs with one byte instructions and add new features.
Some of the one byte instructions had a few issues, mostly related
to the fact that 1 entry tuples should be defined with a comma
at the end such as (0xad,) instead of (0xad).

In addition, functions to assemble the experimental instructions
CALLI, CMPHS, and CMPHU we added to the assembler.

Furthermore, this patch introduces a few callbacks that can greatly
help the programmer write software for this assembler.

One of the things that makes it tedious to program for the vCPU
(defined by ORGs) is the struggle to find a suitable location where
to place the functions. For example, suppose there are 4 functions
f1, f2, f3, and f4 of sizes: 23, 44, 77, and 40 respectively.
In addition, suppose the segments have size 96 bytes (to fit next
to the video RAM location). If one just places the functions in the
natural order of f1, f2, f3 and f4, they will end up using 3 segments:
- segment 1 containing f1 and f2;
- segment 2 containing just f3;
- segment 3 containing just f4;
Note that this is not the optimal order for placing the functions.
Instead, if the functions were placed as in f1, f3, f2, f4, only
2 segments would be used:
- segment 1 containing f1 and f3;
- segment 2 containing f2 and f4;

For that reason, the assembler was extended will a RESOLVE_SEGMENTS()
function, that is used to implement `floating` segments. These floating
segments do not have a fixed start address, and are placed and resolved
by the user callback. Note that callbacks are used in RESOLVE_SEGMENTS()
to allow multiple different placing algorithms.

Lastly, an emit callback was also added. This callback is invoked before
any new instruction is emitted. This can be useful for automatically
creating new segments as soon as one segment becomes full.