Preliminary version of GtForth.
This is an experimental implementation of Forth for the Gigatron TTL
on top of the vCPU. It lacks many features, and is not as fast as
it would be as if it were implemented on top of a native Forth
virtual machine. But it is already capable of executing some basic
words in Forth, such as IF, THEN, BEGIN, WORDS, etc.

This GT1 program should work in all ROM versions since ROMv1, as it
does not use the more recent SYS calls.

Many of the words in GtForth are defined using GtForth itself.
By using the VirtualCpu we can instantiate GtForth and compile
some words in this virtual instance, and save the newly defined
words back into the GT1 file.