Minor hacks to ROM.asm.py to generate listing
Enables the ROM script to be run outside of the repository root,
and be run without arguments. Also makes running it from the tests
easier and faster.

Naturally the running the script without arguments will not generate
a usable image, but I want to be able to see the diff in the listing.