Change how gcl0x.py finds v6502.json
Previously gcl0x.py tried to load Core/v6502.json relative to the
current working directory. This makes it inconvenient to use from
my Contrib area, as I'd need to create a Core directory and copy it
into place (and then always have the right working directory).

The previous behaviour is preserved if Core/v6502.json exists, but
now it will look for the file in the same directory as the
module otherwise. I think this should generally be a safe change for
code that currently works. It may make code that expects to see the
FileNotFoundError fail, but it seems unlikely that anyone is relying
on that.

It's not clear to me exactly why defining these symbols in this module
is actually necessary. I think it may just be to warn if GCL code
defines symbols that clash with the 6502 opcode names.