Demo a function that uses SYS_MultiplyBytes_120
This is a demonstration of SYS_MultiplyBytes_120, and of using
pygtemu for unit testing a GCL function. The conclusion is that it's
not very easy!

Some of the code included should probably end up in pygtemu itself.

I generated a ROM with the relevant code, and it renders the set
in under 8 minutes. This is a significant improvement over
shift-and-add in vCpu, but it might not compare well with a
native shift-and-add approach. I suspect it might be productive to
check if each byte is zero before multiplying, and getting a zero as
a partial product.