Gigatron graph paper
This postscript program will output a 160 x 120 grid on A4 paper.
The output is landscape,
and is marked with the default addresses of the Gigatron display.

It goes to some lengths to get good output
by using the device space units for the grid.

Getting the dashed lines right was a total pain!